Store district_surcharge, registration_fees_waived and instructor_scholarships in DB

+++ db/migrate/20140227064400_add_district_surcharge_to_semester.rb
@@ -1,7 +1,7 @@
 class AddDistrictSurchargeToSemester < ActiveRecord::Migration
   def change
-    add_column :semesters, :district_surcharge, :double
-    add_column :semesters, :registration_fees_waived, :double
-    add_column :semesters, :instructor_scholarships, :double
+    add_column :semesters, :district_surcharge, :float
+    add_column :semesters, :registration_fees_waived, :float
+    add_column :semesters, :instructor_scholarships, :float
   end
 end
